Save work on Optix version

+++ optx/thing.h
@@ -7,7 +7,7 @@ class Optics ;
 
 struct Binding {
 	P p ;
-	double t ;
+	float t ;
 	V    normal ;
 	bool facing ;
 	shared_ptr<Optics> optics ;
@@ -15,7 +15,7 @@ struct Binding {
 
 class Thing {
 	public:
-		virtual bool hit( const Ray& ray, double tmin, double tmax, Binding& binding ) const = 0 ;
+		virtual bool hit( const Ray& ray, float tmin, float tmax, Binding& binding ) const = 0 ;
 } ;
 
 #endif // THING_H
